requirements.txt
Flask==2.3.3
pandas==2.0.3
scikit-learn==1.3.1
numpy==1.26.0
matplotlib==3.8.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
Flask_cors